feat: Add popper layout engine to `Popover` (#1178)
* Add the popper layout engine to `Popover`. This will allow the popover to automatically change

![image](https://user-images.githubusercontent.com/4495057/109476289-d62a2200-7aca-11eb-9eaf-574deb8fc3ca.png)

For the sake of backwards compatibility, internally I created `PopoverLegacy`, which gets used if the `referenceElement` prop has not been included. Keeping this up to date is going to be a problem. I'm tempted to go "No, you must go through every `Popover` component in your codebase, and update the props accordingly". I'm happy to fix all the existing code in perf-ui. LMK if you'd like to go down this path.

Contrary to the Tooltip and Menu components, I did not add a `portalSelector` prop. It will be the consumer's responsibility to wrap the popover in a portal if needed. This is due to a difference in the component design. The Popover doesn't wrap around an element, you pass it the reference.

# Breaking changes

BREAKING CHANGE: To quickly fix all breaking changes, simply import `PopoverLegacy` instead of `Popover` from `@kaizen/draft-popover`. Codemod:

```
codemod -m -d ./src --extensions tsx,jsx 'import \{([\,\w\s]*)Popover([\w\W]*)\} from "@kaizen\/draft-popover"' 'import {\1PopoverLegacy\2} from "@kaizen/draft-popover"'
codemod -m -d ./src --extensions tsx,jsx '<Popover([\w\W]*)<\/Popover>' '<PopoverLegacy\1</PopoverLegacy>'
```

With that said, it is recommended that you start using the `usePopover` hook (see the stories for more examples), or the new `Popover` component if you are using a class component (see the `usePopover` hook to see how this needs to be wired up).

BREAKING CHANGE: The `boxOffset` property has been removed. Now that the new `Popover` uses popper, it will automatically determine the offset. This property was also being overloaded with two types of offsets, depending on the value type, which was confusing.

BREAKING CHANGE: With the new version of `Popover`, the `side` and `position` props are no longer used. Use `placement` instead. Also note, the `placement` describes the position of the popover, relative to the referenceElement, which is unlike the legacy popover, where it describes the placement of the arrow.

BREAKING CHANGE: We now use popper as the positioning engine, which shouldn't cause any breaking change, but please test your popovers when upgrading this package.

# Side updates
* Fix the `automationId` prop not getting added to the root div element
* Remove the `id` prop, which wasn't getting used.