Fix type errors now that the react-select type definitions are too clever
The new type definitions have a new (optional) second generic parameter, `IsMulti`.

The default generic type is `false` - so the only valid value for the `isMulti:IsMulti`
prop is `false`.

You can add a type parameter to make it `boolean` so that `true` or `false` are both accepted.

We need to do this for both Select and AsyncSelect prop definitions