fix: add additional background colors to HeroCard component
Previously the HeroCard component used wisteria700 as a hardcoded background color, this change allows variations to the background color, allowing the new survey summary page to use it and match the current designs.

I thought about adding a transparent background, and then controlling the background color in a child element, but there was too much absolute positioning and/or negative margins involved for it to feel like a clean solution.

This is still not a very extensible way to define variations, but I think a larger conversation needs to happen before dynamic styling is included in any components.