feat: Toast notification manager (#758)
* Pairing pass at a NotificationManager

* Simplify initialization/state

* Position wrapper and bring into line with code style

* Add react-dom as a peerDep

* Split up into cleaner components

* Add specs for NotificationsList

* Set an automation ID on the portal

* Run safely on server

* Update ToastNotification spec

* Add content below InlineNotifications

* Tweak notification animations to be poppier

This also remove the super awkward margin delay on InlineNotifications

* Use <div> in place of <p> to avoid nesting errors

* Change children to message so API is clearer

* Allow ID to be optionally passed

* Tweak toast animation again to collapse after moving

* Make API exposure clearer

* Call render immediate after registering the setter

Means that any tick-delays in React mounting will be resolved when the setter is registered. Before this you could (potentially though not practically) get into a situation where the first call to addToastNotification would not display.

* Add specs for ToastNotificationManager

* Add navigation for context to ToastNotification stories

* Update snapshots

* Separate calling types

Allows us to call addToastNotification with defaults in a type-safe way

* Update stories/docs

* Shift position to align with title as best as possible

In an ideal world we’d anchor from a *measurement* of the titleblock et al in various scenarios, but for now we’re going for "good enough in most situations".

* Tweak position and overflow

This is a pretty weak solution as it hard-crops the list for now. Ideally we take full control of the list and its position manually with JavaScript so we aren’t reliant on implicit positioning and can handle overflow and scrolling more easily.

* Make mobile/desktop display consistent

* Return empty fragment

* Add additional story to demonstrate update-in-place

Co-authored-by: Ally Beaton <allanna.beaton@gmail.com>