fix: update Collapsible component styles
This brings the Collapsible component closer in line with the branding guidelines and uses styles closer to those used in the Card component.

This may not have been done previously because curved corners using border-radius and sticky headers tend not to play well. There is a solution for this in the new `overflow: clip` property that is being implemented in modern browsers.

As a fallback, we can still use `overflow: visible` for older browsers, this has the side effect of children without border radius overlapping their parents, but this is a minor visual glitch that many people won't notice.

Talking to Lindsey, we agreed that this can be considered progressive enhancement. We will prioritize correct behaviour (sticky headers) over correct appearance (border radius).

fixup