feat: add Collapsible configuration for background highlighting
While we can provide custom header JSX to the Collapsible component, we cannot determine if the header should be highlighted when open. One of the designs we are implementing for the survey summary page requires that the highlighting be disabled, so I have added that open.

The highlighting on hover is always enabled as this signals that the header is interactive.