test: clean up title block tests (#1556)
* fix: incorrect prop for onClick

* fix: Clean up logs for title block

This was a little convaluted. Title block has some tests that
re-render, which was causing warnings to fire in the console as
we weren't wrapping calls in `act`. After migrating the tests that
needed that behaviour (using async and waitFor), _more_ warnings
were spat at me, because now that we await behaviour, when we
click links jsDOM does *not* like that.

So I made all the anchor tags with an # href, so that we didn't
navigate. Then I thought I should use the 'props' that are
building these tests in the assertions so that no one has to
manually change each one again.

And _that_ led me to notice that we were testing consts declared
in the describe block across multiple tests. This leads to one
failure in a jest.fn expection. So I moved them all to beforeEach.

This commit cleans things up but does _not_ fix the broken test.

* fix: Change the expectation on the jest.fn

Should this be true?? It's not

Co-authored-by: Ally Beaton <allanna.beaton@gmail.com>