feat: elm tag allow text wrapping option + dismiss button size fix (#1671)
* Elm tag: add missing iconWrapper class to dismiss button

previously the icon was not constrained in size, and was way
too big. Now it's the same size as the react one.

* add allowTextWrapping option to Elm Tag, Select and SelectField

This is a option to deal with tags with user generated content
that is so large that the text escapes the available horizontal
space (such as in a multi-select field). Currently the tag
completely breaks if this happens. There's an existing option
to add truncation (ellipsis) with a fixed max width, but the problem
with that is there is no tooltip, there is no way to distinguish
tags that start with the same text (this is a real problem in production)

Ideally, we'd implement the tooltip when the tag truncates, but this
is unfortunately in the way too hard basket for Elm. You need to measure
the width of the text with an effect, and then the component requires state
to keep track of it - it would be a major change to the component. Not only
would it be difficult to implement but it would be a major headache for consumers.

So, as a quick fix, this option allows the tag's content to grow vertically so
that it can be read, and is in response to an urgent required fix in production.