feat: Stylelint plugin and codemod (for updating deprecated tokens etc) (#1520)
* Many changes to make it cleaner and better:
- Refactor to use postcss-value-parser which unlocks much more potential
- Refactor function transformations to use this
- Code cleanup and readability (but could still use more)
- Lint Imports functionality improved tremendously
- Basic README

* More improvements:
- Linting and autofixing invalid usage of tokens in calc()
- Better reporting of invalid function usages
- More tests
- Option to allow stylelint to autofix

* Cleanup dependencies

* README additions

* Comments

* Cleanup

* Prepare script

* Improve publishability

* Add remove unused imports to CLI

* Fix equation and negated variables linting

* Big refactor to split into multiple rules

* Error message cleanup

* Further error message improvement

* Compile color manipulation functions in place

Use the sass compiler to partially compile stylesheets (Only for SASS at the moment, but it's possible to do the same for less, just requires some refactoring because the compiler doesn't have a "sync" method, and our whole architecture is currently synchronous)

* Make the function compiler work with new tokens

* Improve AtRule linting

* Improve add-alpha function transformation

* Revert package.json version

* Fix TS error

* Remove stylelint config
We can't have a stylelint config here until we publish the package (it seems)

* Fix jest tests

Ignore @kaizen/design-tokens/{sass,less,css} when mapping to identity-object-proxy

* Jest config fix for design-tokens assets

* Multiple improvements:

- No transitive tokens rule, meaning a bunch of new autofixes
- Tests for no-transitive-rokens
- Cleanup code a bit with extra comments
- Autofixing in stylelint should be much more reliable now, due to not using `decl.variable` to determine whether something is a variable.

* Remove stylelint from  stylelint-plugin for now

* No transitive tokens usage fix and readibility

* Further improve no-transitive-tokens

* No transitive tokens improvements

Now resolves variable values lexically.

* Cleanup and improvement on message

* Fix transparentize function

* Improve no-invalid-functions rule
Add a more comprehensive allow-list of functions that can contain CSS variables (a list of all non-compile-time CSS functions based on the spec)

* Improve linting on variables.

Before, the program would just bail if it encountered a variable.
Now, it will continue to find more issues within variable values, but won't autofix.

* Better detection of deprecated tokens

This change happened because variable-identifiers we're causing lint errors.

* Fix lint error

* Improve detection of deprecated tokens in atrule

Also, another attempt at making CI linting work

* Update packages/stylelint-plugin/README.md

Co-authored-by: Jason O'Neil <jason.oneil@gmail.com>

* Parse json tokens and refactor

* Readme improvement and un-expose CLI

* Unify postcss versions and add comments

* README updates and some spec language improvements

* disableFixing option

* More comments

* Revert upgrades of postcss packages

* Remove CLI

* Update codemod

* Fix typos

* Remove PostCSS plugin

* Move containsUnmigratableFunction

* Remove allRulesEnabled

* Amend comment

* Add comment to transformDecl

* Fix replaceTokenInVariable

* Update packages/stylelint-plugin/src/rules/no-invalid-equations.ts

Co-authored-by: Jason O'Neil <jason.oneil@gmail.com>

* Update packages/stylelint-plugin/types/stylelint.d.ts

Co-authored-by: Jason O'Neil <jason.oneil@gmail.com>

* Localise single use constants to invalid func rule

* Un-export function that's unused externally

* no-transitive-tokens comment

* eslintignore and prettierignore improvement

* Test less file cleanup

* Remove sass dependency

The postcss and stylelint packages can parse sass code enough for our
codemod. We only had sass itself to attempt to compile sass code so
we could substitute hardcoded values in place of variables and functions.

Given we no longer want this behaviour, we can remove the dependency

* Change codemod tests to be more agressive in splitting whitespace.

This will allow us to begin laying out our test cases over multiple lines
for easier readability.

* Update tests to no longer have anything about compiling colour functions

(We'll update the code to make the tests pass next)

* No longer compile colour functions to hex values

Remove anything to do with running a sass compiler to substitute inline hex values in place of colour mixing functions.

And also:
- Clean up unused code and unused exports
- No longer export utility functions that aren't used in other modules

* Fix to prevent using css variable tokens in `saturate()`

(saturate is a standard CSS function which would allow CSS variables using the rgb-params format,
but SASS overloads the function and computes at compile time. So we're now warning about any usages
of variable tokens in this function)

Also: add a comment explaining why `declContainsInvalidFunctions` is exported.

* Move `getReplacementForDeprecatedToken` to no-deprecated-tokens module

This utility function isn't used in any other places

* Export a configuration preset rather than the plugin

This makes it easier for us to add new rules and modify defaults in future with less configuration updates required to consuming apps - they can simply write `extends: ["@kaizen/stylelint-plugin"]` to get the plugin and our default rule configuration.

* Add a dev-config.js for running this config in Kaizen itself

(and fix the path in dev-plugin)

We can delete the stylelint package's own `.stylelintrc.json` file once we add the extends configuration to our root config.

* Remove `codemod` file that is unused outside tests

I've moved the two functions into the local spec.

This is theoretically useful later (we'd thought about using a CLI), but I'm trying to reduce the surface area of this package.

* Move the utilities into a utilities folder

(and add a note to explain that it's a bit of a mess)

I've also moved some of the shared types from "types.ts" to a different
place if they were only used once.

* Give the rules easier to understand names

* Refactor to add a `RuleDefinition` type for all of our rule exports

This makes the plugin definition a little cleaner, and the rule definitions themselves
easier to follow.

* Remove `unusedImports` configuration option that is unused

* Tried to improve the rule descriptions in the README

* Improve the comment on dev-plugin.js (matching dev-config.js)

* Remove unused dependencies

* Move dependencies to their correct category

* Fix typo

* Remove instructions that are likely to go out of date

* Improve description for `kaizen/no-tokens-in-variables`

* Refactor equation specs to be clearer about what is or isn't valid

Co-authored-by: Jason O'Neil <jason.oneil@gmail.com>
Co-authored-by: Ally Beaton <allanna.beaton@gmail.com>