feat: Add new Heart color names to design tokens
This adds new colors that are in sync with Heart/the rebrand (e.g. Wisteria -> Purple).
It is part of the bigger plan to completely remove the old ones, which we'll do in a later commit.
The rationale behind having both old and new in the same repo temporarily is that the alternative, pinning a certain version of design tokens in every package which depends on it and doing some release trickery, is a lot more awkward.