delete component-library/styles color,layout & type .scss files
BREAKING CHANGE

The release deletes the `color.scss`, `type.scss` and `layout.scss` files in
`@kaizen/component-library/styles` because they are a duplicate of equivalent files
in the `@kaizen/deprecated-component-library-helpers` package. We want the files
in the other package to be used in favor of component-library scss styles.

Along with this release is a Stylelint rule that you can use to do an automatic upgrade
for this breaking change. All you need to do (with a caveat mentioned below) is:
- upgrade to the latest version of `@kaizen/stylelint-plugin`
- run `yarn stylelint --fix`
- you should see (possible many) changes in scss files where `component-library`
  imports are replaced with ones from `deprecated-component-library-helpers`

*caveats*
- If a scss file contains any `stylelint-ignore` rules at all (including even just one
  `stylelint-ignore-next-line`), the stylelint fix won't be applied anywhere in that file,
  so you will need to manually upgrade it. This is due to a current limitation of Stylelint.
  You will notice this if when you run `stylelint --fix` and some files are not automatically fixed,
  but instead you just get a warning.

Please don't hesitate to reach out to the #prod_design_systems channel on Slack
if you'd like some help with the upgrade - we are more than happy to assist.