feat: forward refs through select component
We should allow refs to be passed through all components, but this is particularly important for elements that we are using in forms. Previously we had not allowed this in the Select component, even though the underlying library `react-select` does allow this.

Because `ref` is handled differently to other props, we need to wrap the components in a ForwardRef function. The forwardRef function provides a mutableRef type and `react-select` expects this to be an immutableRef, so we also have to type the arguments in a way that satisfies `react-select` rather than using generic params in forwardRef.

Because we are wrapping the component with forwardRef, we also need to set the displayName, as React will not be able to infer it. This will help with debugging, ensuring the components do not appear as `forwardRef(Unknown)` in the dev tools.