fix: show tooltips in chromatic tests
The visual regression tests around the tooltip component render the button that triggers the tooltip, but not the tooltip itself. This means that these tests provide little value, unless we can get the tooltip into the test.

The ideal way to do this would be to trigger a hover event in chromatic before the screenshot is captured, but the current chromatic API doesn't support any interaction or events. Instead, we are putting a new prop into the component to signal that the tooltip should be visible on the first render. This seems like it will only be useful for testing, but someone may find a valid use for it, so it hasn't been named obtusely (a name like `chromaticTestVisibility` would send a signal that this is the only valid usage)