fix: Support design-tokens v3 in stylelint plugin
This change improves the no-invalid-use-of-tokens rule to handle both version 2 and 3 of design-tokens, by allowing for the possibility that a token might not exist anymore within it (but still existing in the deprecated tokens list). To do this we had to also slightly modify how `fixAlphaModificationFunctions` worked (as its ability was predicated on Kaizen tokens existing).
We also improve the way we test the stylelint plugin by allowing test cases to supply a `expectedWarningMessages` field, which causes a more specific assertion to made over what warning message texts are reported, rather than just a number of them.