fix: Use the headless/ui transition callbacks in modals
Modify the onOpen/onClose functions to rely on the callbacks from
the transition library. This ensures our label is visible when we
try to focus it.

We can remove the timeout function as it is no longer necessary.

We also need to keep the have the afterLeave code called in the
unmount of the component, as we do not know how consumers will
dispose of the modal.