chore: Remove pointer-events:none CSS on Checkbox and Radio
These were implemented because we found that Radios and Checkboxes weren't working properly in our surveys on touch devices. This recently broke again because the divs were changed to spans and these rules no longer applied.

After a bit of investigation - the checkboxes and radios weren't working properly because of FastClick (which should be removed), so these can just be removed now.