Fix type errors in Storyboard
Previously these were ignoring "metadata" exports from our storybook modules by returning null if the export was not a function, because most of our stories are something like `export const PrimaryButton = () => <Button primary />`

For some reason this changed in TS4, and we have a new way of checking them.

This feels brittle, if a component function happens to have a static "title" prop, it wouldn't work.

Or if a storybook module had an `export default` metadata without a title, it would also break.

Not sure of a better solution.