feat: Delete deprecated OffCanvas and NavigationBar components from component-library
BREAKING CHANGE: There are a few places (mainly in Murmur) that are importing the NavigationBar stylesheet `~@kaizen/component-library/components/NavigationBar/styles` for a `$ca-navigation-bar__height` variable. Though it's not a reliable way of getting the nav height, these can be replaced by the `$layout-navigation-bar-height` from `~@kaizen/design-tokens/sass/layout`

The only usage of the NavigationBar component itself is on the Kaizen site, so that will need to update before this package can update there.