Remove CustomRichTextEditor idea
While we solved one part of this (sending in a custom schema), we didn't tackle how you would actually create custom toolbar controls.

I've made the decision to abandon this idea, at least for now, in favour of focusing on building the simple version, and making it easy for engineers to build a custom RTE through exported primitives and functions.