fix: add missing imports in deprecated-component-library-helpers/styles/type.scss
To be able to upgrade consumers from packages/component-library/styles/type.scss
to packages/deprecated-component-library-helpers/styles/type.scss the two files need
to have the same immports, even if the import isn't actually used in the file that's
importing it. This is because the version of scss we use doesn't really have an import
system. if type.scss imports color.scss, and the consumer imports type.scss, then the
consumer has a implicit dependency on color.scss, so removing color.scss will break the
consumer if it's using any vars from color.scss.