feat: Add control props to FilterMultiSelect Menu Trigger (#2939)
* feat: Add control props to FilterMultiSelect Menu Trigger

`useMenuTriggerState` allows for two more props, `isOpen` and
`onOpenChange`. Add these props to the FilterMultiSelect and pass
them down to the MenuTriggerProvider. This allows a consumer to
control the menu display via their own state and callbacks.

* fix: Update Root component props to read from sub provider props

Ensure the props for the root component are read from the menu
trigger provider props.

Change the interface to a type so that we can omit the children
and pass no extra props. The eslint/no-empty-interface rule blocks
us from keeping it as an interface.

* fix: When controlled, ensure the boolean prop is not undefined

The defaultOpen prop is always set to false rather than undefined,
when we render a controlled version, we must be explicit with the
isOpen prop.

* fix: Remove the default defaultValue for the menu

I think undefined is an important state to pass for these props,
as it's used to decide whether or not the state is controlled or
uncontrolled.

Remove the defaultValue defaulted to false, so that the undefined
value is respected as a state