feat: Add onSearchInputChange prop to filter multi select
For a consumer to implement async functionality, they need access
to the onChange of the search input. With this new callback,
the user can spy on the changes and decide what to do with the
resulting user interaction.

When the user is spying on the onSearchInputChange, the default
filter behaviour of the SelectionProvider is turned off.

The input is not fully controlled by the consumer, so the state
the input holds is still owned locally by the component. This
solution (useEffect) is not appropriate for that requirement and
should be re-implemented if that functionality is needed.