fix: custom button now forwards custom props
When using the `component` prop of a button, we allow the user to
render a custom component via render props. When we do so, we do
not generate `customProps` that the link or regular button allow
for.

Generate the customProps for the custom button, and pass them
through to the custom Component function.