fix: custom button now forwards custom props (#3356)
* fix: custom button now forwards custom props

When using the `component` prop of a button, we allow the user to
render a custom component via render props. When we do so, we do
not generate `customProps` that the link or regular button allow
for.

Generate the customProps for the custom button, and pass them
through to the custom Component function.

* test: Add a test for GenericButton

Test for three top level render conditions:
  * custom component
  * anchor element
  * button element

Test that we pipe through custom props for each condition.