Force update of @react-stately/datepicker to avoid typescript issue
Version 3.3.0 of react-stately/datepicker generated a TS issue, which was handled
in Kaizen with a `@ts-expect-error` comment. But newer releases (3.4.0 onwards) don't
have the problem, and so no error is found.

Because of a problem with Kaizen packaging accidentally shipping TS source files,
consuming projects, which had different versions of this package, got TS errors saying
that the "ts-expect-error" was unnecessary.

There are two suggested fixes I suggest:

1. Remove the `@ts-expect-error` and force a package upgrade to a version which
  won't produce an error. (this commit).
2. Fix Kaizen package publishing so that TS source files are either not included,
  or are included in a separate folder, so that the TS compiler in consuming projects
  doesn't accidentally parse TS source code.

Fixes FEF-435 - https://cultureamp.atlassian.net/browse/FEF-435