Force update of @react-stately/datepicker to avoid typescript issue (#3563)
* Force update of @react-stately/datepicker to avoid typescript issue

Version 3.3.0 of react-stately/datepicker generated a TS issue, which was handled
in Kaizen with a `@ts-expect-error` comment. But newer releases (3.4.0 onwards) don't
have the problem, and so no error is found.

Because of a problem with Kaizen packaging accidentally shipping TS source files,
consuming projects, which had different versions of this package, got TS errors saying
that the "ts-expect-error" was unnecessary.

There are two suggested fixes I suggest:

1. Remove the `@ts-expect-error` and force a package upgrade to a version which
  won't produce an error. (this commit).
2. Fix Kaizen package publishing so that TS source files are either not included,
  or are included in a separate folder, so that the TS compiler in consuming projects
  doesn't accidentally parse TS source code.

Fixes FEF-435 - https://cultureamp.atlassian.net/browse/FEF-435

* Add changeset

* Fix type errors in FilterSelect after `@react-types/shared` upgrade.

In the latest version of `@react-types/shared`, `value` changed from `value: T` to `value: T | null`. See:

https://github.com/adobe/react-spectrum/commit/c3a32cc0a1195799b250dffd98377933ae17e9bd#diff-432a9108dde6f02ef6ec35ec4ce104f8e0fc02591d9f03afd5a23c0e90b36161R163

This might be a breaking change for consumers of FilterSelect (or even other pacakges?)