add classnameOverride prop to InlineNotification
The use case is that we want to be able to remove the `width: 100%` style as
we have a use case where we want the notification to shrink wrap its contents
and use it alongside another element. Ie: an "inline" InlineNotification.

To do so requires making isSubtle a new property of GenericNofication, as
classnameOverride is already being used by InlineNotification for the
isSubtle variant