Add missing prop descriptions for Radio
We just fixed an accessibility bug in performance-review-cycles-ui
that was caused by an engineer who didn't understand the purpose of
the `name` attribute on radio buttons
(and therefore the `name` prop on `Radio` components):

https://github.com/cultureamp/performance-review-cycles-ui/pull/788

To help avoid similar bugs in future,
this adds documentation for the `name` prop
(along with the `value` prop for good measure).

I couldn't get Storybook to infer these descriptions from
my attempts to add JSDoc comments to the component prop types,
so I've added them manually to the Story meta objects.
Working JSDocs would be preferable, though, as it would provide
helpful in-editor descriptions for these props too.