Exclude Select styles from components layer
Select gets its base styles from React Select, which are not in a layer. To override these base styles, Select's own styles must also not be in a layer.

If we ever need to support consuming apps overriding Select styles, we will need to look into how we might output React Select's styles inside of the components layer.