Force white-space: pre-wrap on RichTextContent (#5751)
Kaizen will soon be loaded into a `components` layer in consuming apps using Tailwind 4.
Once this happens, the `white-space: pre-wrap` style on the `RichTextContent` component would be
overridden by the un-layered styles injected into the page by ProseMirror/TipTap at runtime.
To avoid this, this adds `!important` to the `white-space: pre-wrap` style on the
`RichTextContent` component.