Use Tailwind 3 compatible layer name
Tailwind 3 reserves the layer names base, components and utilities. We cannot publish a layer with any of these names without the Tailwind 3 build in consuming apps trying to process those  layers (which causes a build failure).

Once all consumers are upgraded to Tailwind 4 or later, we can rename `kz-components` to `components`.